+++ docs/changelog.rst
@@ -58,6 +58,8 @@ Detailed list of changes
 
 - macOS: When running the default shell, run it via the login program so that calls to ``getlogin()`` work (:iss:`6511`)
 
+- X11: Fix a crash on startup when the iBUS service returns errors and the GLFW_IM_MODULE env var is set to ibus (:iss:`6650`)
+
 
 0.30.0 [2023-09-18]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~